around 660ms

+++ cmd/clock/main.go
@@ -45,7 +45,7 @@ func main() {
 			}
 
 			fmt.Println()
-			err = clock.CycleDays(uint8(numBalls))
+			err = clock.CycleDays(numBalls)
 			if err != nil {
 				fmt.Println("Error:", err)
 			}
@@ -64,7 +64,7 @@ func main() {
 			}
 
 			fmt.Println()
-			err = clock.State(uint8(numBalls), minutes)
+			err = clock.State(numBalls, minutes)
 			if err != nil {
 				fmt.Println("Error:", err)
 			}
